Update __init__.py
This version of the __init__.py file includes the following changes:

It imports the CPI and Inflation classes from the cpi and inflation modules, respectively.

It specifies the CPI and Inflation classes as part of the package's public API using the __all__ variable. This allows users of the economics package to import these classes directly from the package using from economics import CPI or from economics import Inflation.

It does not include the main() function, as this function is not needed in the updated version of the package.