Update cpi.py
This version of the cpi.py file includes the following enhancements:

Error handling: The load() method includes error handling for cases where the API request fails or the data is not in the expected format.

Caching: The load() method includes support for loading data from a local cache file using the pickle module. The save_cache() and clear_cache() methods allow the user to save the current data to a cache file and delete the cache file, respectively.

Additional data sources: The load() method includes support for loading data from the World Bank API or a local cache file. The user can specify the data source to use when creating an instance of the CPI class.

Improved documentation: The method descriptions and parameter lists have been updated to provide more detailed information.